Fix(ng-if): resize based on ng-if elements getting added and removed.
Previously, dynamically added/removed elements would not be recognized by the ui-layout system.  Allow for items to be added/removed from the system and resize as appropriate.
All unit tests pass.  Fix up js-hint.  Fix it up so that ng-if/ng-repeat can be used before or after other
elements.

Fixes #84