Broadcast message when layout has finished loading
Helps setting propers layout when there are initially collapsed containers.
Ui-layout calculates the layout one container at a time.
One way to make sure there's space left for the container is to set it's
'size' or 'min-size'. When the size is not given, ui-layout must calculate
the dimensions first, then notify the scope, which in turn sets the flag
for collapsed containers.